requirements.txt
livekit-agents>=1.2.0
livekit-plugins-deepgram>=1.2.0
livekit-plugins-openai>=1.2.0
livekit-plugins-silero>=1.2.0
aiohttp>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0